print_agent/requirements.txt
flask
pillow
brother_ql
pyusb
pywin32
